Use the general ReflectClassInfo
