+++ source/slang/slang-serialize-reflection.cpp
@@ -5,6 +5,17 @@
 
 namespace Slang {
 
+bool ReflectClassInfo::isSubClassOfSlow(const ThisType& super) const
+{
+    ReflectClassInfo const* info = this;
+    while (info)
+    {
+        if (info == &super)
+            return true;
+        info = info->m_superClass;
+    }
+    return false;
+}
 
 
 } // namespace Slang
